src/main/java/chess/dao/RoomDao.java
package chess.dao;

import java.sql.Connection;
import java.sql.DriverManager;
import java.sql.PreparedStatement;
import java.sql.ResultSet;
import java.sql.SQLException;
import java.util.ArrayList;
import java.util.List;

public class RoomDao {

    private static final String SERVER = "localhost:13306";
    private static final String DATABASE = "chess";
    private static final String OPTION = "?useSSL=false&serverTimezone=UTC";
    private static final String USERNAME = "root";
    private static final String PASSWORD = "root";

    public Connection getConnection() {
        try {
            return DriverManager.getConnection("jdbc:mysql://" + SERVER + "/" + DATABASE + OPTION, USERNAME, PASSWORD);
        } catch (final SQLException e) {
            System.err.println("DB 연결 오류:" + e.getMessage());
            e.printStackTrace();
            return null;
        }
    }

    public void save(final int boardId) {
        final String query = "INSERT INTO room (board_id) VALUES (?)";

        try (final PreparedStatement preparedStatement = getConnection().prepareStatement(query)) {

            preparedStatement.setInt(1, boardId);

            preparedStatement.addBatch();
            preparedStatement.executeBatch();
        } catch (SQLException e) {
            e.printStackTrace();
            throw new RuntimeException("room 저장 실패");
        }
    }

    public List<Integer> findAll() {
        final String query = "SELECT room_id FROM room;";

        List<Integer> roomNumber = new ArrayList<>();

        try (final var connection = getConnection();
             final var preparedStatement = connection.prepareStatement(query)) {

            ResultSet result = preparedStatement.executeQuery();

            while (result.next()) {
                roomNumber.add(result.getInt("room_id"));
            }

        } catch (SQLException e) {
            e.printStackTrace();
        }

        return roomNumber;
    }
}
